# server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cricket Knowledge RAG Chatbot API",
    "version": "1.0.0",
    "description": "API documentation for the Cricket Knowledge RAG (Retrieval-Augmented Generation) Chatbot. This backend parses documents (PDF, DOCX, TXT), chunks them, embeds them with the Gemini API, stores them in MongoDB Atlas, performs Vector Search, manages phone authentication, and provides personalized quizzes with score points."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local Development Server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/quiz/active": {
      "get": {
        "summary": "Get user's active personalized quiz",
        "description": "Retrieves the active generated quiz or triggers batch generation of 10 new questions if user's ask history changed. Reuses cached context chunks.",
        "responses": {
          "200": {
            "description": "Quiz retrieved successfully (answers omitted).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "quizId": { "type": "string" },
                    "questions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "questionText": { "type": "string" },
                          "options": { "type": "array", "items": { "type": "string" } }
                        }
                      }
                    },
                    "cached": { "type": "boolean" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request (e.g. user history empty)."
          }
        }
      }
    },
    "/quiz/submit": {
      "post": {
        "summary": "Submit quiz answers and earn points",
        "description": "Grades the submitted answers on the server side, calculates score (+5 per correct, -1 per wrong), updates cumulative Cric Points, and logs the attempt.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "quizId": { "type": "string" },
                  "answers": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "questionId": { "type": "string" },
                        "selectedOptionIndex": { "type": "integer", "minimum": 0, "maximum": 3 }
                      }
                    }
                  }
                },
                "required": ["quizId", "answers"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Quiz graded successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "attemptId": { "type": "string" },
                    "score": { "type": "integer" },
                    "totalQuestions": { "type": "integer" },
                    "pointsEarned": { "type": "integer" },
                    "cumulativePoints": { "type": "integer" },
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "questionId": { "type": "string" },
                          "questionText": { "type": "string" },
                          "options": { "type": "array", "items": { "type": "string" } },
                          "selectedOptionIndex": { "type": "integer" },
                          "correctOptionIndex": { "type": "integer" },
                          "isCorrect": { "type": "boolean" },
                          "explanation": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/quiz/attempt/{attemptId}": {
      "get": {
        "summary": "Get details of a past quiz attempt",
        "description": "Retrieves the user's detailed options, score, and corrections for a specific past quiz attempt ID.",
        "parameters": [
          {
            "name": "attemptId",
            "in": "path",
            "required": true,
            "description": "ID of the quiz attempt.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detailed attempt details successfully retrieved.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "attemptId": { "type": "string" },
                    "score": { "type": "integer" },
                    "totalQuestions": { "type": "integer" },
                    "pointsEarned": { "type": "integer" },
                    "completedAt": { "type": "string", "format": "date-time" },
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "questionId": { "type": "string" },
                          "questionText": { "type": "string" },
                          "options": { "type": "array", "items": { "type": "string" } },
                          "selectedOptionIndex": { "type": "integer" },
                          "correctOptionIndex": { "type": "integer" },
                          "isCorrect": { "type": "boolean" },
                          "explanation": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (attempt belongs to another user)."
          },
          "404": {
            "description": "Quiz attempt or associated content not found."
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}